RoutingService refactoring
 - move classes related with routing to separate subdirectory inslide libosmscout
 - modify data structures used for routing to support "multi-database" routing
 - split RoutingSevice to AbstractRoutingService, SimpleRoutingService and new MultiDBRoutingService
