+++ libosmscout-import/include/osmscout/import/GenWayWayDat.h
@@ -29,7 +29,7 @@
 #include <osmscout/Coord.h>
 #include <osmscout/CoordDataFile.h>
 #include <osmscout/NumericIndex.h>
-#include <osmscout/TurnRestriction.h>
+#include <osmscout/routing/TurnRestriction.h>
 
 #include <osmscout/import/Import.h>
 #include <osmscout/import/RawWay.h>
