+++ libosmscout/include/Makefile.am
@@ -77,6 +77,7 @@ nobase_include_HEADERS= osmscout/private/Config.h \
                         osmscout/SRTM.h \
                         osmscout/LocationService.h \
                         osmscout/POIService.h \
+                        osmscout/AbstractRoutingService.h \
                         osmscout/RoutingService.h
 
 if OSMSCOUT_HAVE_SSE2
