+++ libosmscout/include/meson.build
@@ -64,7 +64,9 @@ osmscoutHeader = [
             'osmscout/RouteNode.h',
             'osmscout/RoutePostprocessor.h',
             'osmscout/RoutingProfile.h',
+            'osmscout/AbstractRoutingService.h',
             'osmscout/RoutingService.h',
+            'osmscout/MultiDBRouting.h',
             'osmscout/SRTM.h',
             'osmscout/Tag.h',
             'osmscout/TurnRestriction.h',
