+++ libosmscout/include/osmscout/routing/MultiDBRoutingService.h
@@ -1,0 +1,182 @@
+#ifndef OSMSCOUT_MULTIDBROUTINGSERVICE_H
+#define OSMSCOUT_MULTIDBROUTINGSERVICE_H
+
+/*
+  This source is part of the libosmscout library
+  Copyright (C) 2016  Tim Teulings
+  Copyright (C) 2017  Lukas Karas
+
+  This library is free software; you can redistribute it and/or
+  modify it under the terms of the GNU Lesser General Public
+  License as published by the Free Software Foundation; either
+  version 2.1 of the License, or (at your option) any later version.
+
+  This library is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+  Lesser General Public License for more details.
+
+  You should have received a copy of the GNU Lesser General Public
+  License along with this library; if not, write to the Free Software
+  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
+*/
+
+#include <osmscout/Pixel.h>
+#include <osmscout/routing/AbstractRoutingService.h>
+#include <osmscout/routing/SimpleRoutingService.h>
+
+#include "DBFileOffset.h"
+
+namespace osmscout {
+
+  /**
+   * \ingroup Routing
+   *
+   * Helper container for MultiDBRoutingService
+   */
+  class RouterDBFiles{
+  public:
+    IndexedDataFile<Id,RouteNode>       routeNodeDataFile;     //!< Cached access to the 'route.dat' file
+    IndexedDataFile<Id,Intersection>    junctionDataFile;      //!< Cached access to the 'junctions.dat' file
+    ObjectVariantDataFile               objectVariantDataFile;
+  public:
+    RouterDBFiles();
+    virtual ~RouterDBFiles();
+    bool Open(DatabaseRef database);
+    void Close();
+  };
+
+  typedef std::shared_ptr<RouterDBFiles> RouterDBFilesRef;
+
+  /**
+   * \ingroup Routing
+   *
+   * Util class for routing cross databases
+   */
+  class OSMSCOUT_API MultiDBRoutingService CLASS_FINAL : public AbstractRoutingService<MultiDBRoutingState>
+  {
+  private:
+
+  public:
+    typedef std::function<RoutingProfileRef(const DatabaseRef&)> RoutingProfileBuilder;
+
+  private:
+    static const double CELL_MAGNIFICATION;
+    static const double LAT_CELL_FACTOR;
+    static const double LON_CELL_FACTOR;
+
+  private:
+    std::map<std::string,DatabaseId>              databaseMap;
+    // TODO: replace with std::vector indexed by DatabaseId
+    std::map<DatabaseId,DatabaseRef>              databases;
+    std::map<DatabaseId,SimpleRoutingServiceRef>  services;
+    std::map<DatabaseId,RoutingProfileRef>        profiles;
+
+    std::map<DatabaseId,RouterDBFilesRef>         routerFiles;
+
+    bool  isOpen;
+
+  private:
+    virtual Vehicle GetVehicle(const MultiDBRoutingState& state);
+
+    virtual bool CanUseForward(const MultiDBRoutingState& state,
+                               const DatabaseId& database,
+                               const WayRef& way);
+
+    virtual bool CanUseBackward(const MultiDBRoutingState& state,
+                                const DatabaseId& database,
+                                const WayRef& way);
+
+    virtual double GetCosts(const MultiDBRoutingState& state,
+                            const DatabaseId database,
+                            const RouteNode& routeNode,
+                            size_t pathIndex);
+
+    virtual double GetCosts(const MultiDBRoutingState& state,
+                            const DatabaseId database,
+                            const WayRef &way,
+                            double wayLength);
+
+    virtual double GetEstimateCosts(const MultiDBRoutingState& state,
+                                    const DatabaseId database,
+                                    double targetDistance);
+
+    virtual double GetCostLimit(const MultiDBRoutingState& state,
+                                const DatabaseId database,
+                                double targetDistance);
+
+    virtual bool GetRouteNodesByOffset(const std::set<DBFileOffset> &routeNodeOffsets,
+                                       std::unordered_map<DBFileOffset,RouteNodeRef> &routeNodeMap);
+
+    virtual bool GetRouteNodeByOffset(const DBFileOffset &offset,
+                                      RouteNodeRef &node);
+
+    virtual bool GetRouteNodeOffset(const DatabaseId &database,
+                                    const Id &id,
+                                    FileOffset &offset);
+
+    virtual bool GetWayByOffset(const DBFileOffset &offset,
+                                WayRef &way);
+
+    virtual bool GetWaysByOffset(const std::set<DBFileOffset> &wayOffsets,
+                                 std::unordered_map<DBFileOffset,WayRef> &wayMap);
+
+    virtual bool GetAreasByOffset(const std::set<DBFileOffset> &areaOffsets,
+                                  std::unordered_map<DBFileOffset,AreaRef> &areaMap);
+
+    /*
+    virtual void ResolveRNodeChainToList(DBFileOffset finalRouteNode,
+                                         const ClosedSet& closedSet,
+                                         std::list<VNode>& nodes);
+     */
+
+    virtual bool ResolveRouteDataJunctions(RouteData& route);
+
+    virtual bool GetRouteNode(const DatabaseId &database,
+                              const Id &id,
+                              RouteNodeRef &node);
+    
+    Pixel GetCell(const osmscout::GeoCoord& coord);
+    
+    bool ReadCellsForRoutingTree(osmscout::Database& database,
+                                 std::unordered_set<uint64_t>& cells);
+
+    bool ReadRouteNodesForCells(osmscout::Database& database,
+                                std::unordered_set<uint64_t>& cells,
+                                std::unordered_set<osmscout::Id>& routeNodes);
+
+    bool FindCommonRoutingNodes(DatabaseRef &database1,
+                                DatabaseRef &database2,
+                                std::set<Id> &commonRouteNodes);
+    
+    bool ReadRouteNodeEntries(DatabaseRef &database,
+                              std::set<Id> &commonRouteNodes,
+                              std::unordered_map<osmscout::Id,osmscout::RouteNodeRef> &routeNodeEntries);
+
+    virtual bool CanUse(const MultiDBRoutingState& state,
+                        const DatabaseId database,
+                        const RouteNode& routeNode,
+                        size_t pathIndex);
+
+  public:
+    MultiDBRoutingService(const RouterParameter& parameter,
+                          std::vector<DatabaseRef> databases);
+
+    virtual ~MultiDBRoutingService();
+
+    bool Open(RoutingProfileBuilder routingProfileBuilder);
+    
+    void Close();
+
+    RoutePosition GetClosestRoutableNode(const GeoCoord& coord,
+                                         double radius=1000,
+                                         std::string databasePathHint="") const;
+
+    RoutingResult CalculateRoute(const RoutePosition &start,
+                                 const RoutePosition &target,
+                                 const RoutingParameter &parameter);
+  };
+}
+
+#endif /* MULTIDBROUTING_H */
+
