+++ libosmscout/include/osmscout/routing/Route.h
