+++ libosmscout/include/osmscout/routing/RouteNode.h
