+++ libosmscout/include/osmscout/routing/TurnRestriction.h
