+++ libosmscout/src/osmscout/routing/Route.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
 */
 
-#include <osmscout/Route.h>
+#include <osmscout/routing/Route.h>
 
 #include <sstream>
