+++ libosmscout/src/osmscout/routing/RouteNode.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
 */
 
-#include <osmscout/RouteNode.h>
+#include <osmscout/routing/RouteNode.h>
 
 #include <limits>
