+++ libosmscout/src/osmscout/routing/TurnRestriction.cpp
@@ -17,7 +17,7 @@
   Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307  USA
 */
 
-#include <osmscout/TurnRestriction.h>
+#include <osmscout/routing/TurnRestriction.h>
 
 namespace osmscout {
